moved messageGet & bulk & typing indicator.
